refactor: move download zip to utils
